+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - [AWS Lambda] Add trace correlation support for Lambda invocations triggered by SQS messages (including SNS-to-SQS).
+- Add trace continuity support for SNS-to-SQS bridge for SQS tracing.
 
 ## 1.127.0
 - AWS SDK v3: Instrumentation of DynamoDB